requirements.txt
numpy
scipy
torch
torchvision
opencv-python
yacs
tqdm
pillow
matplotlib
sklearn
scikit-image
svgpathtools

cairosvg
wand
svglib
reportlab